chose: classe de servico renomeada
- ProdutoServico renomeada para ProdutoInsercaoServico
